Limit import cleanup to numerics and target info
